// src/app/shared/types/authState.interface.ts
import { BackendErrorsInterface } from "src/app/shared/types/backendErrors.interface";
import { CurrentUserInterface } from "src/app/shared/types/currentUser.interface";

export interface AuthStateInterface {
    isSubmitting: boolean
    currentUser: CurrentUserInterface | null
    isLogged : boolean | null
    validationErrors: BackendErrorsInterface | null
    isLoading: boolean
}